requirements.txt
psycopg2==2.8.6
numpy==1.19.2
streamlit==0.80.0
pandas==1.1.3
mysql_connector_repackaged==0.3.1
Pillow==8.3.0
